Modified itoa to accept intmax, patched edge case for itoa_base

+++ ft_itoa.c
@@ -6,13 +6,13 @@
 /*   By: dslogrov <dslogrove@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/22 14:23:38 by dslogrov          #+#    #+#             */
-/*   Updated: 2018/05/22 14:36:53 by dslogrov         ###   ########.fr       */
+/*   Updated: 2018/06/18 14:15:50 by dslogrov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_itoa(int n)
+char	*ft_itoa(intmax_t n)
 {
 	return (ft_itoa_base(n, 10));
 }
